Solve day-24 with a RB-Tree as map for the field
The map has the advantage of beeing dynamic (while still having good
lookup times in contrast to a list) in contrast to an array
which has a static size -> tiles structure couls grow out of bounds.

Nevertheless the Tree/Map has a pourer performance in comparison with
the array on a simple lookup.

Storing x/y min/max in the tree is a necessity arising from a non static
datastructure to be able to print it and to be able to do the
white/black-rule in part2.

The fact that x/y min/max never shrink (in absolute) might be a problem
if the whiteRule applies too often, but the only way I can think of
would be iterating over the whole Tree for searching the new
minima/maxima which isn't that nice either.